+++ src/Lykke.Service.SwiftCredentials/Models/EmailRequestCommand.cs
@@ -4,7 +4,7 @@
     {
         public string ClientId { set; get; }
         public string RegulationId { set; get; }
-        public double Amount { set; get; }
+        public decimal Amount { set; get; }
         public string AssetId { set; get; }
     }
 }
